pyproject.toml
[project]
name = "langchain-intro-and-memory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.8",
    "jupyter>=1.1.1",
    "langchain-core>=0.3.74",
    "langchain[groq]>=0.3.27",
    "langgraph>=0.6.4",
    "python-dotenv>=1.1.1",
]
